Delete license.txt
This project is dual licensed. The server is GPL (later it turns into GPLv2 [will try to track down the discussion there] this GPLv3 must of been added in error.) For this project as of 11/2020 we are clearly putting the license file in each subfolder so there is no confusion what license belongs to what code/files.