Link to File
Dual license repo of The Lacuna Expance GPLv2 and MIT serated for clarity using submodules.

The GPL has been for the server code and MIT for the web based client.

The GPL license is found in /info/ folder. It was moved from the root folder to /info/ folder on this commit:
 - https://github.com/plainblack/Lacuna-Server-Open/blob/17d22cbb60288d2c27c41427c6d5ef1d7ba1bb55/info/LICENSE
 - https://github.com/communityus/info/blob/main/license.txt
 - Starting from commit 30 we will begin closely tracking the changes. /bin/ /docs/ /lib/Lacuna/ /t/ /var/ are impacted. The rest of the folders contain no content aside from the /var/www/public/ folder which contains non-copyleft content that is later added to using MIT from the Kenó Antigen project. A GPL license is added to the project by @haarg 9 years ago.
 1. https://github.com/plainblack/Lacuna-Server-Open/tree/2e5164f7eb8f51102a0ca28e3ae32880248a95ed
 
Our repo may go forwards and backwards in time for archival purposes and for feature comparisons. To be clear the GPL license can always be easily traced in the main repro from Plainblack Corp.
- https://github.com/plainblack/Lacuna-Server-Open